+++ NEWS.md
@@ -1,5 +1,10 @@
 # excelPlot 0.2.1
 
+* Catch case of path = "." in plotExcelFolder
+* Add FLAGopenExcel and FLAGtemp to plotExcel directly.
+
+# excelPlot 0.2.1
+
 * Add diffing capability. plotExcelFolder now also has compareToCommit argument to compare diffing natively.
 
 # excelPlot 0.2.0